Remove passing in of whether or not a screen should be deleted.
UI now controls whether screen can be deleted by disabling the button.
Before, a boolean was passed down through the structure to the reducers,
but it is not confined to the UI, so the delete button is disabled in the
components panel when the screen is selected.  Screens can only be deleted
by clicking the delete button in the screens panel.